+++ src/pypnnomenclature/migrations/data/nomenclatures.sql
@@ -7,11 +7,12 @@ SET client_min_messages = warning;
 
 CREATE SCHEMA ref_nomenclatures;
 
-SET search_path = ref_nomenclatures, pg_catalog;
+SET search_path = ref_nomenclatures, pg_catalog, public;
 
 -------------
 --FUNCTIONS--
 -------------
+
 
 CREATE OR REPLACE FUNCTION get_id_nomenclature_type(mytype character varying) RETURNS integer
 IMMUTABLE